refactor: move styling to css file, file system++
