+++ fresh.gen.ts
@@ -3,18 +3,18 @@
 // This file is automatically updated during development when running `dev.ts`.
 
 import config from "./deno.json" assert { type: "json" };
-import * as $0 from "./routes/[name].tsx";
-import * as $1 from "./routes/api/joke.ts";
-import * as $2 from "./routes/index.tsx";
-import * as $3 from "./routes/this.tsx";
+import * as $0 from "./routes/[blog]/this.tsx";
+import * as $1 from "./routes/[name].tsx";
+import * as $2 from "./routes/api/joke.ts";
+import * as $3 from "./routes/index.tsx";
 import * as $$0 from "./islands/ThemeSwitcher.tsx";
 
 const manifest = {
   routes: {
-    "./routes/[name].tsx": $0,
-    "./routes/api/joke.ts": $1,
-    "./routes/index.tsx": $2,
-    "./routes/this.tsx": $3,
+    "./routes/[blog]/this.tsx": $0,
+    "./routes/[name].tsx": $1,
+    "./routes/api/joke.ts": $2,
+    "./routes/index.tsx": $3,
   },
   islands: {
     "./islands/ThemeSwitcher.tsx": $$0,
